Move to using index.js for fmg-core and fmg-counting-game.
