+++ packages/fmg-core/index.js
@@ -1,0 +1,3 @@
+export { State } from './src/state';
+export { Channel } from './src/channel';
+export { toHex32, padBytes32 } from './src/utils';
