+++ packages/fmg-core/src/state.js
@@ -1,34 +1,4 @@
 import { toHex32, padBytes32 } from './utils';
-import { soliditySha3 } from 'web3-utils';
-
-class Channel {
-  constructor(channelType, channelNonce, participants) {
-    this.channelType = channelType;
-    this.channelNonce = channelNonce;
-    this.participants = participants;
-  }
-
-  get numberOfParticipants() {
-    return this.participants.length;
-  }
-
-  get id() {
-    return soliditySha3(
-      { type: 'address', value: this.channelType },
-      { type: 'uint256', value: this.channelNonce },
-      { type: 'address[]', value: this.participants },
-    );
-  }
-
-  toHex() {
-    return (
-      padBytes32(this.channelType) +
-      toHex32(this.channelNonce).substr(2) +
-      toHex32(this.numberOfParticipants).substr(2) +
-      this.participants.map(x => padBytes32(x).substr(2)).join("")
-    )
-  }
-}
 
 class State {
   constructor({channel, stateType, turnNum, resolution, stateCount}) {
@@ -69,4 +39,4 @@ State.StateTypes = {
 
 Object.freeze(State.StateTypes);
 
-export { Channel, State };
+export { State };
