+++ packages/fmg-counting-game/package.json
@@ -2,7 +2,7 @@
   "name": "fmg-counting-game",
   "version": "0.0.1",
   "description": "Counting game",
-  "main": "index.js",
+  "main": "src/index.js",
   "directories": {
     "doc": "docs",
     "test": "test"
@@ -14,6 +14,7 @@
   "author": "Tom Close",
   "license": "MIT",
   "dependencies": {
+    "fmg-core": "^0.0.1",
     "npm": "^5.7.1",
     "sha3": "^1.2.2",
     "web3-utils": "^1.0.0-beta.30"
